Indexing functionality and improved error handling
Two additional properties will be read out of scss.json, useIndex and
indexFiilePath.

If useIndex is set to true, then all of the scss or sass files in the
project will be copied as imports into one scss file. By default will
be index.scss in the root of the project, or can be specified with the
indexFilePath attribute. This file is the only scss file which will be
compiled.

The error that renderSync throws should be an object but is a string
for some reason. This is now being parsed into an object and passing
better information to the compile step error

Changed some formatting for clarity